LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/parse-error-location.tentative-expected.txt

CONSOLE MESSAGE: SyntaxError: JSON Parse error: Expected '}'

FAIL JSON modules: parse error file location assert_equals: expected "http://localhost:8800/html/semantics/scripting-1/the-script-element/json-module/parse-error.json" but got ""
